% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mask.R
\name{create_mask}
\alias{create_mask}
\title{Create a Mask}
\usage{
create_mask(bb, prj, interactive = FALSE, add = FALSE)
}
\arguments{
\item{bb}{either a bounding box (bbox object), a numeric vector (xmin, ymin, 
xmax, ymax) or an sf POLYGON or MULTIPOLYGON object.}

\item{prj}{a projection string}

\item{interactive}{define the mask interactively}

\item{add}{add the mask to the current plot}
}
\value{
An sf object
}
\description{
Create a mask based on a bounding box, an sf POLYGON or 
MULTIPOLYGON or an interactively defined rectancle.
}
