% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inset_rbindr.R
\name{inset_rbinder}
\alias{inset_rbinder}
\title{Rbind Spatial*DataFrame Objects With Rows and Cols Handling}
\usage{
inset_rbinder(l = list())
}
\arguments{
\item{l}{a list of POLYGON or MULTIPOLYGON objects}
}
\value{
An sf POLYGON or MULTIPOLYGON object
}
\description{
Take a list of sf POLYGON or MULTIPOLYGON objects and output a 
single sf POLYGON or MULTIPOLYGON object.
}
