% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{mapinsetr}
\alias{mapinsetr}
\title{mapinsetr Package}
\description{
Map insets are small zoom-in or zoom-out maps that focus on particular 
territories. \code{mapinsetr} provides a set of functions that helps to 
create such insets.
}
