% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inset_rbindr.R
\name{inset_rbinder}
\alias{inset_rbinder}
\title{Rbind Spatial*DataFrame Objects With Rows and Cols Handling}
\usage{
inset_rbinder(l = list())
}
\arguments{
\item{l}{a list of Spatial*DataFrame}
}
\value{
A Spatial*DataFrame
}
\description{
Take a list of Spatial*DataFrame and output a single Spatial*DataFrame
}
\examples{
if(require(cartography)){
  data(nuts2006)
  mybbox <- rbind(c(4235059,4507390),c(2559828, 2784798)) 
  mymask <- create_mask(bb = mybbox, prj = proj4string(nuts0.spdf), add=FALSE)
  inset <- move_and_resize(spdf = nuts0.spdf, mask = mymask, 
                           xy = c(5566624, 3599815), k = 5)
  newnuts <- inset_rbinder(list(nuts0.spdf, inset))  
  plot(newnuts)
}
}

