% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_r.R
\name{m_r}
\alias{m_r}
\title{Move and Resize in Box}
\usage{
m_r(x, mask, y, verbose = FALSE)
}
\arguments{
\item{x}{the layer to cut, resize and move}

\item{mask}{the targeted area in x}

\item{y}{destination}

\item{verbose}{display k}
}
\value{
a layer
}
\description{
do stuff
}
\examples{
library(sf)
library(mapsf)
nc <- st_read(system.file("shape/nc.shp", package="sf"))
# Ne fonctionne qu'avec 2 fonds projetés
nc <- st_transform(nc, 32119)
# Créer des boites
y <- st_as_sfc(st_bbox(
  c(xmin = 270000, ymin = 00000,
    xmax = 550000, ymax = 100000),
  crs = 32119))
y <- st_make_grid(y, n = c(3,2))
mf_map(nc)
mf_map(y, add = TRUE)
# plusieurs objets agrandis
inset1 <- m_r(x = nc, mask = nc[1,], y = y[1])
mf_map(nc[1, ], col = 2, add = TRUE)
mf_map(inset1, col = 2, add = TRUE)
# plusieurs objets agrandis
inset2 <- m_r(x = nc, mask = nc[2,], y = y[2])
mf_map(nc[2, ], col = 3, add = TRUE)
mf_map(inset2, col = 3, add = TRUE)
# un seul objet, réduit
inset3 <- m_r(x = nc[nc$CNTY_ID == 2000, ], 
              mask = nc[nc$CNTY_ID == 2000, ], 
              y = y[3])
mf_map(nc[nc$CNTY_ID == 2000, ], col = 4, add = TRUE)
mf_map(inset3, col = 4, add = TRUE)
# plusieurs objets dans une autre proj
mtq <- mf_get_mtq()
mvd <- m_r(x = mtq, mask = mtq, y = y[4])
mf_map(mvd, col = 5, add = TRUE)
}
