% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_and_resize.R
\name{move_and_resize}
\alias{move_and_resize}
\title{Move and Resize an sf Object}
\usage{
move_and_resize(x, mask = NULL, xy, prj, k = 1)
}
\arguments{
\item{x}{an sf POINT, POLYON or MULTIPOLYGON object to resize and move.}

\item{mask}{an sf or sfc POLYGON or MULTIPOLYGON object used to select the 
area to move an resize.}

\item{xy}{coordinates used to move the inset, bottomleft corner of the 
inset.}

\item{prj}{CRS string of the output projection of the inset.}

\item{k}{factor used to resize.}
}
\value{
An sf object is returned.
}
\description{
Move and resize a simple feature collection of polygons, 
multipolygons or points.
}
\examples{
library(sf)
nc <- st_read(system.file("shape/nc.shp", package="sf"))
nc <- st_transform(nc, 32119)

plot(st_geometry(nc))
bb <- st_bbox(nc[nc$CNTY_ID == '2030',])
mask <- create_mask(bb = bb, add = TRUE)
inset <- move_and_resize(nc, mask, xy = c(190000, 1000), k = 3)
plot(st_geometry(inset), add = TRUE)

plot(st_geometry(nc))
mask <- st_buffer(st_centroid(nc[nc$CNTY_ID == 2026,]),dist = 30000)
plot(st_geometry(mask), border = "red", lwd = 2, add = TRUE)
inset <- move_and_resize(nc, mask, xy = c(270000, 5000), k = 2.5)
plot(st_geometry(inset), add = TRUE)
}
