% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_and_resize.R
\name{move_and_resize}
\alias{move_and_resize}
\title{Move and Resize a Spatial*DataFrame}
\usage{
move_and_resize(x, mask = NULL, xy, prj, k = 1)
}
\arguments{
\item{x}{an sf POLYON or MULTIPOLYGON object to resize and move.}

\item{mask}{an sf POLYON or MULTIPOLYGON object used to select the area to move an resize.}

\item{xy}{coordinates used for the move process, bottomleft corner of the inset.}

\item{prj}{outputted projection of the inset.}

\item{k}{factor used to resize.}
}
\value{
An sf object
}
\description{
Move and Resize an sf POLYON or MULTIPOLYGON object
}
